CPPlot - Use int to iterate over the bests
It is easier to read and don't generate warnings

+++ src/Charts/CPPlot.cpp
@@ -605,7 +605,7 @@ CPPlot::plotModel()
                 pdModel->setMinutes(false);
                 QVector<double> power(static_cast<int>(pdModel->size()));
                 QVector<double> wprime(static_cast<int>(pdModel->size()));
-                for (size_t t = 0; t < pdModel->size(); t++) {
+                for (int t = 0; t < pdModel->size(); t++) {
                     power[t] = pdModel->y(t+1);
                     wprime[t] = (pdModel->y(t+1)-veloCP) * (pdModel->x(t+1)); // Joules
                 }
@@ -2258,9 +2258,9 @@ CPPlot::exportBests(QString filename)
     foreach(QwtPlotCurve *bestsCurve, bestsCurves) {
 
         // just output for the bests curve
-        for (size_t i=0; i<bestsCurve->data()->size(); i++) {
-            const double xvalue = bestsCurve->sample(static_cast<int>(i)).x();
-            const double yvalue = bestsCurve->sample(static_cast<int>(i)).y();
+        for (int i=0; i<bestsCurve->data()->size(); i++) {
+            const double xvalue = bestsCurve->sample(i).x();
+            const double yvalue = bestsCurve->sample(i).y();
             const double modelvalue = expmodel ? pdModel->y(xvalue) : 0;
 
             int index = xvalue * 60.00f;
